feat: add All users card to admin dashboard
