+++ src/app/(DashboardLayout)/dashboard/page.tsx
@@ -3,7 +3,7 @@ import { Grid, Box } from '@mui/material';
 import PageContainer from '../components/container/PageContainer';
 // components
 import SalesOverview from '../components/dashboard/UserRegistration';
-import YearlyBreakup from '../components/dashboard/YearlyBreakup';
+import YearlyBreakup from '../components/dashboard/AllUsers';
 import RecentTransactions from '../components/dashboard/RecentTransactions';
 import ProductPerformance from '../components/dashboard/ProductPerformance';
 import Blog from '../components/dashboard/Blog';
